Crate darkwikid

Modules

jsonrpc ğŸ”’

lcs ğŸ”’

patch ğŸ”’

util ğŸ”’

Structs

Args ğŸ”’

DarkWiki ğŸ”’

DarkWiki object

DarkWikiSettings ğŸ”’

Settings struct used to hold some metadata for DarkWiki

WORKSPACES ğŸ”’

This is where we hold our workspaces, so we are also able to refresh them on SIGHUP.

Constants

CONFIG_FILE

CONFIG_FILE_CONTENTS

LOCAL_ID_PATH ğŸ”’

SYNC_ID_PATH ğŸ”’

Functions

handle_signals ğŸ”’

main ğŸ”’

realmain ğŸ”’

Type Definitions

Patches ğŸ”’