Module darkwikid::util

Functions

decrypt_patch

Decrypt a patch using a NaCl crypto_secretbox given an EncryptedPatch and a Key.

encrypt_patch

Encrypt a patch using a NaCl crypto_secretbox given a Patch and a Key.

get_docs_paths

TODO: DOCUMENT ME FIXME: There’s checking of file extensions here. Take care that the rest of the code is robust against this attack.

parse_b58_secret 🔒

Parse a base58 string for a crypto_secretbox secret.

parse_workspaces

Parse a TOML string for configured workspaces and return an HashMap of parsed data. Does not error on failures, just warns if something is misconfigured.

path_to_id

Hash a path and workspace, and encode with base58, providing an ID.

str_to_chars

Split a &str into a vector of each of its chars.